Start rafactoring and moving windows native to repo.
Ok, so re-refactoring. :-(
And there's more to come but get this in while it works . . .

New terminology to better fit a mental model of what's happening.
  pages -> views
  browser -> application

First (working) addition of 'doze code.
Specs drive notepad and my own app.
There's a whole buch of starting appium etc.
And some wrapping of a bunch of things.
And the sample app sucks but it sorta works.

Make string element #get work with WinAppDriver.
It gets text as #text instead of #attribute(:value)

Change naming/calling convention for find eleemnt.
For now we try the first selector an element has.